api/requirements.txt
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter
django-extensions 
pygraphviz
django-countries
django-cors-headers